requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pydantic-settings==2.0.3
python-dotenv==1.0.0
prometheus-client==0.19.0
httpx==0.25.1
pytest==7.4.3
pytest-asyncio==0.21.1
structlog==24.1.0
python-json-logger==2.0.7
python-logstash==0.4.8
requests==2.31.0
opentelemetry-api==1.31.1
opentelemetry-sdk==1.31.1
opentelemetry-instrumentation==0.52b1
opentelemetry-instrumentation-asgi==0.52b1
opentelemetry-instrumentation-fastapi==0.52b1
opentelemetry-exporter-jaeger==1.21.0
opentelemetry-exporter-jaeger-proto-grpc==1.21.0
opentelemetry-exporter-jaeger-thrift==1.21.0
opentelemetry-semantic-conventions==0.52b1
opentelemetry-util-http==0.52b1
locust==2.24.0 